Add files via upload
These photos are the results from the "SHBD iNaturalist Assignment" code.  I created two plots using the 'ggplot' command line. 

One visualizes the occurrences of the distribution of the white-tailed deer (Odocoileus virginianus), with 'bio1' (or Mean Annual Temperature) as the background over the United States.

Another visualizes the distribution of the white-tailed deer (Odocoileus virginianus) as a histogram compared with 'bio1' (or Mean Annual Temperature).